Move diff specific functionality into diff/helpers
