+++ vessel/diff/diff_command.py
@@ -35,21 +35,21 @@ from typing import Any, Optional
 import yaml
 
 from vessel.diff.helpers import metadata_diff
-from vessel.diff.helpers.failure import FailureSummary
-from vessel.diff.helpers.file_diff import FileDiffs
-from vessel.diff.helpers.flag import Flag
-from vessel.diff.helpers.metadata_diff import MetadataDiffs
-from vessel.utils import oci, umoci
-from vessel.utils.checksum import (
+from vessel.diff.helpers.checksum import (
     generate_filesummary_and_checksum,
     hash_folder_contents,
     load_checksum_metadata,
     write_checksum_metadata,
 )
-from vessel.utils.diffoscope import (
+from vessel.diff.helpers.diffoscope import (
     DiffoscopeParser,
     build_diffoscope_command,
 )
+from vessel.diff.helpers.failure import FailureSummary
+from vessel.diff.helpers.file_diff import FileDiffs
+from vessel.diff.helpers.flag import Flag
+from vessel.diff.helpers.metadata_diff import MetadataDiffs
+from vessel.utils import oci, umoci
 from vessel.utils.skopeo import skopeo_copy
 from vessel.utils.uri import ImageURI
